requirement.txt
adal==1.2.2
alabaster==0.7.10
aldryn-boilerplates==0.8.0
altgraph==0.16.1
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.9
argcomplete==1.10.2
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
awsebcli==3.15.3
azure-batch==8.0.0
azure-cli-command-modules-nspkg==2.0.3
azure-cli-core==2.0.76
azure-cli-nspkg==3.0.4
azure-cli-telemetry==1.0.4
azure-common==1.1.23
azure-cosmos==3.1.2
azure-datalake-store==0.0.48
azure-graphrbac==0.60.0
azure-keyvault==1.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.1.0
azure-mgmt-appconfiguration==0.3.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.52.0
azure-mgmt-batch==7.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==5.0.0
azure-mgmt-compute==8.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc7
azure-mgmt-containerservice==7.0.0
azure-mgmt-cosmosdb==0.8.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==2.2.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hdinsight==1.1.0
azure-mgmt-imagebuilder==0.2.1
azure-mgmt-iotcentral==1.0.0
azure-mgmt-iothub==0.8.2
azure-mgmt-keyvault==1.1.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==1.1.1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.6.0
azure-mgmt-network==7.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.3.1
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.4.0
azure-mgmt-redis==6.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.3.1
azure-mgmt-resource==4.0.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.1.0
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.3.0
azure-mgmt-sql==0.15.0
azure-mgmt-sqlvirtualmachine==0.4.0
azure-mgmt-storage==5.0.0
azure-mgmt-trafficmanager==0.51.0
azure-multiapi-storage==0.2.4
azure-nspkg==3.0.2
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
botocore==1.12.253
Bottleneck==1.2.1
branca==0.3.1
CacheControl==0.12.5
cement==2.8.2
certifi==2019.9.11
cffi==1.10.0
chardet==3.0.3
Click==7.0
cloudpickle==0.2.2
clyent==1.2.2
cms==0.0.2
colorama==0.4.1
comtypes==1.1.2
conda==4.3.30
contextlib2==0.5.5
coverage==4.5.4
crispy-forms-bulma-django2==1.1.5
crispy-forms-foundation-demo==0.5.0
crispy-forms-propeller==0.0.2a10
crispy-forms-semantic-ui==0.2.0
crispy-semantic-ui==1.0.2
crispy-templates==0.2.0
crispys-webkit==0.1.141
cryptography==2.8
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
decorator==4.0.11
distributed==1.16.3
Django==2.2.6
django-admin-honeypot==1.1.0
django-appconf==1.0.3
django-axes==5.0.18
Django-Bootstrap3-Validator==0.3.3
django-braces==1.13.0
django-commentator==0.0.4
django-contrib-comments==1.9.1
django-crispy-bootstrap==0.1.1.1
django-crispy-bulma==0.1.2
django-crispy-forms==1.7.2
django-crispy-forms-fancy-formsets==0.1
django-crispy-forms-ng==2.0.0
django-cryptography==0.3
django-defender==0.6.2
django-drift-chatwidget==0.1.1
django-filter==2.2.0
django-generic-aggregation==0.4.0
django-haystack==2.8.1
django-ipware==2.1.0
django-js-asset==1.2.2
git+git://github.com/adilshehzad786/Ryp-form.git#egg=django-machina
django-markdown2==0.3.1
django-model-utils==3.2.0
django-mptt==0.10.0
django-ranged-response==0.2.0
django-ratelimit-backend==2.0
django-session-security==2.6.5
django-simple-captcha==0.5.12
django-simple-history==2.7.3
django-simple-ratings==0.3.3
django-simple-search==1.0.2
django-slds-crispyforms==0.1.0
django-sslserver==0.21
django-star==0.999
django-star-ratings==0.8.0
django-storages==1.8
django-tables2-bootstrap4==2.0.0b5
django-taggit==1.1.0
django-widget-tweaks==1.4.5
djangocms-comments==0.2.3
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
fabric==2.5.0
fastcache==1.0.2
Flask==0.12.2
Flask-Cors==3.0.2
folium==0.10.0
future==0.16.0
gevent==1.2.1
greenlet==0.4.12
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.999
humanfriendly==4.18
idna==2.5
imagesize==0.7.1
invoke==1.3.0
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.0
isort==4.2.5
itsdangerous==0.24
javaproperties==0.5.1
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.4
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
knack==0.6.3
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==3.1.1
markdown2==2.3.8
markdown2social==0.3
MarkupSafe==1.1.1
matplotlib==2.0.2
menuinst==1.4.7
mistune==0.7.4
mock==2.0.0
mpmath==0.19
msgpack==0.6.2
msgpack-python==0.4.8
msrest==0.6.10
msrestazure==0.6.2
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
ngrok==0.0.1
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
oauthlib==3.1.0
odo==0.5.0
olefile==0.44
opencv-python==4.1.1.26
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
paramiko==2.6.0
partd==0.3.8
path.py==10.3.1
pathlib2==2.2.1
pathspec==0.5.9
patsy==0.4.1
pbr==5.4.3
pefile==2019.4.18
pep8==1.7.0
pickleshare==0.7.4
Pillow==5.0.0
Pillow-PIL==0.1.dev0
pinax-comments==1.0.3
pinax-ratings==3.0.3
pip-download==0.0.6
pip-downloader==0.2
pip-freezer==0.2.0
ply==3.10
portalocker==1.5.2
prompt-toolkit==1.0.14
psutil==5.6.5
py==1.4.33
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pydocumentdb==2.3.5
pyflakes==1.5.0
Pygments==2.4.2
PyInstaller==3.5
PyJWT==1.7.1
pykismet3==0.1.1
pylint==1.6.4
PyNaCl==1.3.0
pynput==1.4.4
pyodbc==4.0.16
pyOpenSSL==19.1.0
pyparsing==2.1.4
pypiwin32==223
pyreadline==2.1
pytest==3.0.7
python-dateutil==2.6.0
python-http-client==3.2.1
python-markdown-comments==0.0.0
python-nmap==0.6.1
python-slugify==4.0.0
pytz==2019.1
PyWavelets==0.5.2
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.1.2
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
scp==0.13.2
seaborn==0.7.1
semantic-version==2.5.0
sendgrid==3.6.5
sendgrid-django==4.2.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
sqlparse==0.3.0
sshtunnel==0.1.5
statsmodels==0.8.0
swapper==1.1.1
sympy==1.0
tables==3.2.2
tabulate==0.8.6
tblib==1.3.2
termcolor==1.1.0
testpath==0.3
text-unidecode==1.3
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
ua-parser==0.8.0
unicodecsv==0.14.1
UNKNOWN==0.0.0
urljoin==1.0.0
urllib3==1.24.3
urlparser==0.1.2
user-agents==2.0
virtualenv==16.7.5
vsts-cd-manager==1.0.2
waitress==1.3.1
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.12.2
Whoosh==2.7.4
widgetsnbextension==2.0.0
win-unicode-console==0.5
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
xmltodict==0.12.0
YURL==1.0.0
zict==0.1.2
